CLDC-4077: Expire review app images after 30 days (from 14)
review apps are staying open for longer now as we're in support so the 14 day limit was becoming difficult to manage.

we cannot remove this limit as we need some sort of expiration policy though AWS options are limited to last n days or last n images. last n images could cause deploying one review app to break another which is not useful